Refactor some pubsub-related code in the conn
For subscriptions/unsubscriptions, I enqueued a single item in the queue
of the genserver with all the channels of the subscriptions. Then, when
a "subscribe" or "unsubscribe" message arrived, I dequeued, removed a
channel, and enqueued again until there were no channels left. I
simplified the code by just enqueuing a {:subscribe, ...} message for
each channel (same for unsubscribe) so that it can just be popped out of
the queue.