Fix a concurrency bug with pubsub
Ok, this is what was happening. Whenever a connection received
a (p)subscribe call, it set state.pubsub to true. When new TCP data
arrived, the connection reacted differently based on the above-mentioned
state.pubsub: if true, it would handle a pubsub message, otherwise a
regular message.

This caused problems when the connection received messages in an order
like this:

{:commands, ...} -> connection
{:subscribe, ...} -> connection
{:tcp_data, ...} -> connection

The {:tcp_data, ...} message had to be the response to the first
commands (as Redis guarantees that order), but it was treated as a
pubsub message since the {:subscribe, ...} message arrived before the
TCP one.

This has now been fixed by changing the handling of TCP messages when
state.pubsub is true: we first check the queue for clients waiting to
get a response to commands, and if there are we treat the TCP message as
not a pubsub message yet.