Don't remove the password from the conn state
Before this commit, the password in the `opts` field of the state of the
connection was replaced with `:redacted` if auth was successful so that
we avoided printing it in case of failures (when GenServers print the
state). This was obviously causing problems when trying to reconnect
with a password, as the original password was now unavailable.

In this commit, I just removed this "feature".