Fix Redix.stop/1 to be synchronous and not leave zombie processes
Redix.Connection starts two GenServers internally (both with
start_link). However, these GenServers are not trapping exists, and thus they do
not exit when Redix.stop/1 is called, since that makes the Redix.Connection
process exit with reason :normal. With this commit, I changed Redix.stop/1 to be
synchronous and use GenServer.stop/3 (instead of stopping through a returned
:stop tuple), and then I implemented the terminate/2 callback for the
Redix.Connection process: in this callback, I peacefully stop the two child
GenServers (if the terminate reason is :normal) so that they're never orphaned
because either the link will bring them down or we will manually.