Add sentinel support for older versions of Redis
As part of connecting to a sentinel, we need to verify that the server
we are told is the primary actually thinks it is the primary. This is
usually done via the ROLE command.

Unfortunately, the ROLE command was introduced in version 2.8.12 so when
Redix attempts to verify the role of the primary, it gets an error
saying the server doesn't understand ROLE.

According to the Redis docs, if the ROLE command isn't available, we
can resort to calling "INFO replication" and parsing out the "role" -
which is what this commit does.