Unify Telemtry events about connection established
Before this commit, we were emitting a [:redix, :reconnection] Telemetry
event when a connection (both normal as well as pub/sub) established a
connection to Redis after a disconnection. With this commit, we're
removing that event and effectively introducing the [:redix,
:connection] event, which is also emitted when the connection
establishes successfully the first time. This event carries a
:reconnection key in the metadata: this key is a boolean telling whether
the connection is after a reconnection or not. This means that the
functiona behavior of the logging, for example, is left completely
unchanged.

I also added some tests and updated the documentation.