Support Redis 6 ACL (#189)
Redix currently ignores the username if specified in the connection URI.
For versions of Redis before v6, this was correct, because Redis had no
concept of usernames.

Redis v6, however, has the new ACL system (https://redis.io/topics/acl)
which allows you to authenticate as a given user.

This commit adds a new ":username" option to the allowed start options,
and parses this value from a URI if given. If the username and password
are both present, it then uses the new two argument form of the "AUTH"
command when connecting to Redis: "AUTH <username> <password>".
The old one argument form ("AUTH <password>") is used if a password is
given, but no username is given.