Fall back to no ACL if Redis doesn't support it
Closes #198.

In previous commits, we started using "AUTH username password" if a username was
provided, since Redis 6.0.0 started supporting ACL. However, this broke
earlier Redix versions because Redix used to ignore the username in
URIs (as a documented behavior).

With this commit, we still issue "AUTH username password" but we fall
back to "AUTH password" if Redis explicitly errors out with "wrong
number of arguments".