Match empty string path on database connection URI
Elixir 1.13 has changed the behaviour of URI.parse/1 to default
the path to an empty string instead of nil.

https://github.com/elixir-lang/elixir/commit/e414cc0e83e2b389638ec04149ce761599b5d149

In elixir 1.12.3 the path is nil

```
URI.parse("redis://localhost:6379")
%URI{
  authority: "localhost:6379",
  fragment: nil,
  host: "localhost",
  path: nil,
  port: 6379,
  query: nil,
  scheme: "redis",
  userinfo: nil
}
```

In elixir 1.13-rc.0 the path is an empty string

```
URI.parse("redis://localhost:6379")
%URI{
  authority: "localhost:6379",
  fragment: nil,
  host: "localhost",
  path: "",
  port: 6379,
  query: nil,
  scheme: "redis",
  userinfo: nil
}
```

This fix adds an empty string to the nil database path match.