Handle Redis servers that disable CLIENT command
See https://github.com/whatyouhide/redix/issues/192.

Before this commit, we would not handle the case where the Redis server
had disabled CLIENT commands. In that case, when using noreply_*
functions (which issue CLIENT commands under the hood), we would mess up
the state of the ETS table used for queueing commands to reply to. That
would show up as a hard-to-debug ArgumentError because of an unhandled
:"$end_of_table" return value from :ets.first/1.

With this commit, we made a couple of fixes to:

  * Handle empty ETS request queues
  * Handle errors returned by "CLIENT REPLY OFF" in
    Redix.noreply_pipeline/3

We also added a new Docker Compose Redis instance to test against a
server that disables CLIENT commands.