Fix telemetry warning
warning from logs:

```

The function passed as a handler with ID "redix-default-telemetry-handler" is a local function.

This means that it is either an anonymous function or a capture of a function without a module specified. That may cause a performance penalty when calling that handler. For more details see the note in `telemetry:attach/4` documentation.

https://hexdocs.pm/telemetry/telemetry.html#attach/4

```