Default to using Erlang certificate store (#279)
If supplied, the `:cacerts` option overrides `:cacertfile`, so we have
to check if the user has specified a different CA certificate store
before adding them to the default options.

The OTP team no longer supports Erlang versions earlier than 25+, so we
are assuming that `:public_key.cacerts_get/0` is available.

Closes https://github.com/whatyouhide/redix/issues/278